Refactor User.java and UserInfo.java into a single module
