+++ src/main/java/com/syncstream_dev/syncstream/model/User.java
@@ -1,23 +1,33 @@
-package com.syncstream_dev.syncstream.model.user;
+package com.syncstream_dev.syncstream.model;
 
 import java.time.LocalDateTime;
 
 
 public class User {
-    private UserInfo userInfo;
     private String userId;
     private String username;
     private int messagesSent;
     private int points;
+    private String token;
+    private String email;
+    private char[] password;
+    private LocalDateTime creationDate;
+    private LocalDateTime lastLogin;
+    private String avatarUrl;
 
     // Register User Constructor
     public User(String userId, String username, String token, String email, String password, String avatarUrl) {
         // TODO: Database integration
-        this.userInfo = new UserInfo(token, email, password, LocalDateTime.now(), LocalDateTime.now(), avatarUrl);
         this.userId = userId;
         this.username = username;
         this.messagesSent = 0;
         this.points = 0;
+        this.token = token;
+        this.email = email;
+        this.password = password.toCharArray();
+        this.creationDate = LocalDateTime.now();
+        this.lastLogin = LocalDateTime.now();
+        this.avatarUrl = avatarUrl;
     }
 
     // Login User Constructor
@@ -27,10 +37,6 @@ public class User {
     }
 
     // Getters
-    public String getUserToken() {
-        return userInfo.getToken();
-    }
-
     public String getUserId() {
         return userId;
     }
@@ -45,6 +51,30 @@ public class User {
 
     public int getPoints() {
         return points;
+    }
+
+    public String getToken() {
+        return token;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public char[] getPassword() {
+        return password;
+    }
+
+    public LocalDateTime getCreationDate() {
+        return creationDate;
+    }
+
+    public LocalDateTime getLastLogin() {
+        return lastLogin;
+    }
+
+    public String getAvatarUrl() {
+        return avatarUrl;
     }
 
     // Overrides
