+++ src/main/java/com/syncstream_dev/syncstream/model/message/UserMessage.java
@@ -1,6 +1,6 @@
 package com.syncstream_dev.syncstream.model.message;
 
-import com.syncstream_dev.syncstream.model.user.User;
+import com.syncstream_dev.syncstream.model.User;
 
 public class UserMessage extends Message {
     private User sender;
