+++ src/test/java/com/syncstream_dev/syncstream/datamodeltest/messagetest/UserMessageTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import com.syncstream_dev.syncstream.model.message.UserMessage;
-import com.syncstream_dev.syncstream.model.user.User;
+import com.syncstream_dev.syncstream.model.User;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
